feat: 리뷰 1페이지 섹션화 (최상단 UI 렌더링용으로 초기 세팅)
